refactor: use barrel export in order to use mantine components in RSC, refactor breadcrumbs in Header to an RSC
