+++ www/admin/app/dashboard/pages/_components/AddPageModal.tsx
@@ -1,6 +1,6 @@
 import { type ChangeEvent, useState } from "react";
 import { type SubmitHandler, useForm } from "react-hook-form";
-import { ActionIcon, Alert, Group, Modal, Stack, TextInput } from "@mantine/core";
+import { ActionIcon, Alert, Group, Modal, Stack, TextInput } from "@admin/src/components";
 import {
 	DocumentPlusIcon,
 	ExclamationCircleIcon,
