+++ www/admin/app/dashboard/pages/_components/PageTree.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useMemo, useReducer, useState } from "react";
-import { Loader, LoadingOverlay, useMantineColorScheme } from "@mantine/core";
+import { Loader, LoadingOverlay, useMantineColorScheme } from "@admin/src/components";
 import { Section } from "@admin/app/dashboard/_components/Section";
 import { trpcReact } from "@admin/src/utils/trpcReact";
 import type { Page as PageType } from "api/prisma/types";
