+++ www/admin/app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 import { cookies } from "next/headers";
-import type { ColorScheme } from "@mantine/core";
+import type { ColorScheme } from "@admin/src/components";
 import Mantine from "./_providers/mantine";
 import TrpcProvider from "./_providers/trpcProvider";
 import "@admin/src/styles/globals.css";
